v2 service

+++ app/api/user/service.go
@@ -1,29 +1,32 @@
-package service
+package user
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/korableg/V8I.Manager/app/api/user"
-	"github.com/korableg/V8I.Manager/app/api/user/repository"
+
 	"golang.org/x/crypto/bcrypt"
+)
+
+const (
+	adminID int64 = 1
 )
 
 type (
 	Service interface {
-		Add(ctx context.Context, u user.AddUserRequest) (int64, error)
-		Get(ctx context.Context, ID int64) (user.User, error)
-		GetList(ctx context.Context) ([]user.User, error)
-		Update(ctx context.Context, u user.UpdateUserRequest) error
+		Add(ctx context.Context, u AddUserRequest) (int64, error)
+		Get(ctx context.Context, ID int64) (User, error)
+		GetList(ctx context.Context) ([]User, error)
+		Update(ctx context.Context, u UpdateUserRequest) error
 		Delete(ctx context.Context, ID int64) error
 	}
 
 	service struct {
-		userRepo repository.Repository
+		userRepo Repository
 	}
 )
 
-func NewService(userRepo repository.Repository) (*service, error) {
+func NewService(userRepo Repository) (*service, error) {
 	if userRepo == nil {
 		return nil, errors.New("user repository is not defined")
 	}
@@ -35,13 +38,13 @@ func NewService(userRepo repository.Repository) (*service, error) {
 	return s, nil
 }
 
-func (s *service) Add(ctx context.Context, uReq user.AddUserRequest) (int64, error) {
+func (s *service) Add(ctx context.Context, uReq AddUserRequest) (int64, error) {
 	passwordHash, err := hashPassword(uReq.Password)
 	if err != nil {
 		return 0, fmt.Errorf("hashPassword: %w", err)
 	}
 
-	u := user.User{
+	u := User{
 		Name:         uReq.Name,
 		PasswordHash: passwordHash,
 		Token:        uReq.Token,
@@ -56,16 +59,16 @@ func (s *service) Add(ctx context.Context, uReq user.AddUserRequest) (int64, err
 	return id, nil
 }
 
-func (s *service) Get(ctx context.Context, ID int64) (user.User, error) {
-	u, err := s.Get(ctx, ID)
+func (s *service) Get(ctx context.Context, ID int64) (User, error) {
+	u, err := s.userRepo.Get(ctx, ID)
 	if err != nil {
-		return user.User{}, fmt.Errorf("get from store: %w", err)
+		return User{}, fmt.Errorf("get from store: %w", err)
 	}
 
 	return u, nil
 }
 
-func (s *service) GetList(ctx context.Context) ([]user.User, error) {
+func (s *service) GetList(ctx context.Context) ([]User, error) {
 	users, err := s.userRepo.GetList(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("get from store: %w", err)
@@ -74,8 +77,8 @@ func (s *service) GetList(ctx context.Context) ([]user.User, error) {
 	return users, nil
 }
 
-func (s *service) Update(ctx context.Context, uReq user.UpdateUserRequest) error {
-	u := user.User{
+func (s *service) Update(ctx context.Context, uReq UpdateUserRequest) error {
+	u := User{
 		ID:    uReq.ID,
 		Name:  uReq.Name,
 		Token: uReq.Token,
@@ -90,6 +93,10 @@ func (s *service) Update(ctx context.Context, uReq user.UpdateUserRequest) error
 }
 
 func (s service) Delete(ctx context.Context, ID int64) error {
+	if ID == adminID {
+		return errors.New("it is not possible to delete the admin account")
+	}
+
 	if err := s.userRepo.Delete(ctx, ID); err != nil {
 		return fmt.Errorf("delete user from store: %w", err)
 	}
